master
Problem:
The plugin did not support adding the artifact URL from the plugin settings and building a path to a specific artifact based on the commit tag.

Solution/changes:
The plugin has been updated to accept an artifact URL, which is used in conjunction with the build tag to build the URL for the artifact, which is appended to existing notes added to the release.

Notes: